updated to main.js
