+++ src/index.html
@@ -136,7 +136,7 @@
     <script src="./GUI/LoadingScreen.js"></script>
 
 
-    <script type="module" src="./test.js"></script>
+    <script type="module" src="./main.js"></script>
 </body>
 
 </html>
