--- src/test.html
@@ -1,32 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha3/dist/css/bootstrap.min.css" rel="stylesheet"
-        integrity="sha384-KK94CHFLLe+nY2dmCWGMq91rCGa5gtU4mk92HdvYe+M/SXH301p5ILy+dN9+nJOZ" crossorigin="anonymous">
-
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha2/dist/js/bootstrap.bundle.min.js"
-        integrity="sha384-qKXV1j0HvMUeCBQ+QVp7JcfGl760yU08IQ+GpUo5hlbpg51QRiuqHAJz8+BrxE/N"
-        crossorigin="anonymous"></script>
-
-</head>
-
-<body>
-    <div class="accordion accordion-flush" id="accordionFlushExample">
-        <div class="accordion-item">
-            <h2 class="accordion-header" id="flush-IRIS"><button class="accordion-button collapsed" type="button"
-                    data-bs-toggle="collapse" data-bs-target="#flush-collapseIRIS" aria-expanded="false"
-                    aria-controls="flush-collapseIRIS">IRIS</button></h2>
-            <div id="flush-collapseIRIS" class="accordion-collapse collapse" aria-labelledby="flush-headingIRIS"
-                data-bs-parent="#accordionFlushExample">
-                <div class="accordion-body" style="width:100%;"><canvas id="pop"></canvas></div>
-            </div>
-        </div>
-    </div>
-</body>
-
-</html>
